Modified file structure for seed data
